Add cleaned up private version of MapSwipe
Several commits ago, we removed all of the open code from this repository. Early
on, the core developers and supporters of the MapSwipe application wanted to
make it open source, but because of various reasons like credentials and
branding in the codebase, the main application code was kept private.

As volunteers worked to improve the private codebase, the public one fell behind
without much maintenance. Because we ultimately felt like what made MapSwipe
important was it's role in providing a window for volunteers to help our
supporting organizations, we decided to clean up the private repository in a
number of ways:

* We removed sensitive information from this, things that shouldn't be shared
  with the public
* We implemented a number of fixes that removed the need to keep a large portion
  of imported JavaScript modules commited/vendored in the codebase
* We clarified instructions and added additional information for people who wish
  to contribute to the project
* We added some additional resources and scripts to help newcomers get started
  with the application

As with any volunteer effort, one is always left with a feeling that you could
have done more. However, we just know that we have done **good** and will
continue to do **good**.