Install react-native-community/react-native-netinfo
This is part of the Lean Core transition in RN, where some modules
are gradually extracted from the core RN library. It doesn't change functionality
but is needed to prevent problems with the next upgrades of RN where
these bits of code will be removed.

See https://github.com/react-native-community/react-native-netinfo