Split CI workflows into 3 files
This is to allow rerunning ios or android build/deploy
separately in case of failure, as it is not possible with
the previous layout of jobs in a single workflow.
This commit does not change any active code, simply moves
it to new files and adds comments, so it should function
exactly like before.